rename elem restriction kernel file
