+++ backends/magma/kernels/common/elem_restriction.h
